requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.1.0
Django==4.2.9
django-cors-headers==3.13.0
django-dynamic-fixture==4.0.1
django-environ==0.11.2
django-mail-templated==2.6.5
django-mkdocs==0.1.2
django-phonenumber-field==7.0.0
django-rest-swagger==2.2.0
django-sites==0.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
ghp-import==2.1.0
idna==3.6
importlib-metadata==7.0.1
itypes==1.2.0
Jinja2==3.1.3
kombu==5.3.5
Markdown==3.5.2
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.5.3
openapi-codec==1.3.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3.post1
PyYAML==6.0.1
pyyaml_env_tag==0.1
redis==5.0.1
requests==2.31.0
sendgrid==6.9.7
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
watchdog==4.0.0
wcwidth==0.2.13
zipp==3.17.0
